refactor: 상세페이지 리팩터링
