+++ src/app/(route)/news/_components/newsGnb/NewsDetailGnb.tsx
@@ -8,11 +8,12 @@ import { ReportType } from "@/services/board/types/report";
 import { FeedbackContentType } from "@/app/(route)/customer/_types/FeedbackItemType";
 import { useAdminRole } from "@/app/(route)/customer/_utils/adminChecker";
 import SignInModalPopUp from "@/app/_components/SignInModalPopUp";
+import { FeedbackInfoType } from "@/app/(route)/customer/(route)/feedback/_types/FeedbackInfoType";
 
 interface NewsDetailGnbProps {
   title: string;
   type?: "news" | "notice" | "feedback";
-  data?: FeedbackContentType;
+  data?: FeedbackContentType | FeedbackInfoType;
   id?: number;
 }
